refactor(ui): Extract the form payload builder
